again6

+++ 0x00-hello_world/6-size.c
@@ -12,11 +12,11 @@ int main(void)
 	double m;
 	float f;
 
-	printf("Size of char: %lu. byte(s)\n", sizeof(c));
-	printf("Size of an int: %lu.  byte(s)\n", sizeof(i));
-	printf("Size of a long int: %lu.  byte(s)\n", sizeof(l));
-	printf("Size of a long_long int: %lu. byte(s)\n", sizeof(m));
-	pritnf("Size of float: %lu. byte(s)\n", sizeof(f));
+printf("Size of char: %lu. byte(s)\n",(unsigned long)sizeof(c));
+printf("Size of an int: %lu.  byte(s)\n",(unsigned long)sizeof(i));
+printf("Size of a long int: %lu.  byte(s)\n",(unsigned long)sizeof(l));
+printf("Size of a long_long int: %lu. byte(s)\n",(unsigned long)sizeof(m));
+pritnf("Size of float: %lu. byte(s)\n", (unsigned long)sizeof(f));
 
-	return (0);
+return (0);
 }
